feat: start a stopped queue
Add the ability to start a queue that has been stopped by adding a new "stopping" state. You will be able to start a queue that is in the stopped state, but not the consuming or stopping state.

The old consumers are removed when  stopped so will need to be re-added after you've started the queue again.